requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
idna==2.8
mtnmomo==3.0.1
phonenumbers==8.10.20
python-dotenv==0.10.3
requests==2.21.0
urllib3==1.24.3
